requirements.txt
# ATO RAG Pipeline
requests>=2.31
beautifulsoup4>=4.12
sentence-transformers>=2.2
faiss-cpu>=1.7
numpy>=1.24
fastapi>=0.104
uvicorn>=0.24
pydantic>=2.0
